Added Seek and cleand up hashing
Implemented io.Seeker on Tree so that trees are more file like.
Cleaned up hashing using the new GetDigest to take both children
of a branch at the same time.

+++ tree.go
@@ -18,7 +18,7 @@ type Tree struct {
 	leaves         uint32
 	f              *Forest
 	lastBlockLen   uint16
-	pos            int
+	pos            int64
 	complete       bool
 	leavesComplete []bool
 }
